requirements.txt
django==3.2.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==0.21.0
